Add files via upload
When you run this file in a code editor such as VSCode, it will download the highest resolution version of the video to the specified location. PS: Note that downloading videos from YouTube may violate their terms of service, so make sure you have the necessary rights or permissions to download and use the content.