Add files via upload
This script uses the speedtest-cli library to test the download and upload speeds. The speeds are displayed in megabits per second (Mbps). When you run this script, it will initiate the speed test and print out the results.
The accuracy of the speed test results can depend on various factors, including the network conditions and the Speedtest.net server being used for the test.